Rename component from Tabs to NavTabs
