+++ ui/src/core/assets/themes/index.ts
@@ -46,7 +46,7 @@ import { dark as darkCircleMatcher } from './circleMatcher';
 import { dark as darkCircleSegmentation } from './circleSegmentation';
 import { dark as darkCheckbox } from './checkbox';
 import { dark as darkCircleGroupMetrics } from './circleGroupMetrics';
-import { dark as darkTabs } from './tabs';
+import { dark as darkNavTabs } from './navTabs';
 import { dark as darkSummary } from './summary';
 import { zIndex } from '../zindex';
 
@@ -92,7 +92,7 @@ const dark = {
   circleSegmentation: darkCircleSegmentation,
   checkbox: darkCheckbox,
   circleGroupMetrics: darkCircleGroupMetrics,
-  tabs: darkTabs,
+  navTabs: darkNavTabs,
   summary: darkSummary
 };
