+++ ui/src/core/assets/themes/navTabs.ts
